Implement voting inside the app
Instead of depending on an external app to do voting
which is pretty core to the function of this app,
the voting has been internalized. This has the added
benefit of avoiding the inefficient joins necessary
with generic foreign keys and some of the templatetag
gymnastics required to use a generic voting app.